+++ packages/zod-openapi/test/index.test.ts
@@ -1,9 +1,13 @@
 import type { RouteConfig } from '@asteasolutions/zod-to-openapi'
 import type { Context, TypedResponse } from 'hono'
+import { bearerAuth } from 'hono/bearer-auth'
 import { hc } from 'hono/client'
-import { describe, it, expect, expectTypeOf } from 'vitest'
-import { OpenAPIHono, createRoute, z, RouteConfigToTypedResponse } from '../src/index'
-import { Expect, Equal } from 'hono/utils/types'
+import { describe, expect, expectTypeOf, it } from 'vitest'
+import type { RouteConfigToTypedResponse } from '../src/index'
+import { OpenAPIHono, createRoute, z } from '../src/index'
+import type { Equal, Expect } from 'hono/utils/types'
+import type { ServerErrorStatusCode } from 'hono/utils/http-status'
+import { stringify } from 'yaml'
 
 describe('Constructor', () => {
   it('Should not require init object', () => {
@@ -200,11 +204,19 @@ describe('Basic - params', () => {
             responses: {
               '200': {
                 description: 'Get the user',
-                content: { 'application/json': { schema: { $ref: '#/components/schemas/User' } } },
+                content: {
+                  'application/json': {
+                    schema: { $ref: '#/components/schemas/User' },
+                  },
+                },
               },
               '400': {
                 description: 'Error!',
-                content: { 'application/json': { schema: { $ref: '#/components/schemas/Error' } } },
+                content: {
+                  'application/json': {
+                    schema: { $ref: '#/components/schemas/Error' },
+                  },
+                },
               },
             },
           },
@@ -861,6 +873,13 @@ describe('basePath()', () => {
 
     expect(client.api.message.$url().pathname).toBe('/api/message')
   })
+
+  it('Should add the base path to paths', async () => {
+    const res = await app.request('/api/doc')
+    expect(res.status).toBe(200)
+    const data = (await res.json()) as any
+    expect(Object.keys(data.paths)[0]).toBe('/api/message')
+  })
 })
 
 describe('With hc', () => {
@@ -1007,14 +1026,13 @@ describe('With hc', () => {
       },
       (result, c) => {
         if (!result.success) {
-          const res = c.json(
+          return c.json(
             {
               ok: false,
               source: 'routeHook' as const,
             },
             400
           )
-          return res
         }
       }
     )
@@ -1465,6 +1483,14 @@ describe('RouteConfigToTypedResponse', () => {
           },
           description: 'Error!',
         },
+        '5XX': {
+          content: {
+            'application/json': {
+              schema: ErrorSchema,
+            },
+          },
+          description: 'Server Error!',
+        },
       },
     }
 
@@ -1486,6 +1512,55 @@ describe('RouteConfigToTypedResponse', () => {
           400,
           'json'
         >
+      | TypedResponse<
+          {
+            ok: boolean
+          },
+          ServerErrorStatusCode,
+          'json'
+        >
     type verify = Expect<Equal<Expected, Actual>>
+  })
+})
+
+describe('Generate YAML', () => {
+  it('Should generate YAML with Middleware', async () => {
+    const app = new OpenAPIHono()
+    app.openapi(
+      createRoute({
+        method: 'get',
+        path: '/books',
+        middleware: [
+          bearerAuth({
+            verifyToken: (_, __) => {
+              return true
+            },
+          }),
+        ],
+        responses: {
+          200: {
+            description: 'Books',
+            content: {
+              'application/json': {
+                schema: z.array(
+                  z.object({
+                    title: z.string(),
+                  })
+                ),
+              },
+            },
+          },
+        },
+      }),
+      (c) => c.json([{ title: 'foo' }])
+    )
+    const doc = app.getOpenAPI31Document({
+      openapi: '3.1.0',
+      info: {
+        title: 'My API',
+        version: '1.0.0',
+      },
+    })
+    expect(() => stringify(doc)).to.not.throw()
   })
 })